Dataset_Agent/agents/dataset/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
supabase==2.9.0
kaggle==1.6.6
google-cloud-storage==2.14.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0
